(PC-19910): refactor(share): move WebShareModal in features/share
