+++ src/features/share/pages/WebShareModal.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent } from 'react'
 
-import { WebShareModalProps } from './types'
+import { WebShareModalProps } from 'features/share/types'
 
 export const WebShareModal: FunctionComponent<WebShareModalProps> = () => null
