+++ src/features/venue/components/VenueHeader/VenueHeader.tsx
@@ -5,10 +5,10 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
+import { WebShareModal } from 'features/share/pages/WebShareModal'
 import { useShareVenue } from 'features/venue/helpers/useShareVenue'
 import { analytics } from 'libs/firebase/analytics'
 import { useWhiteStatusBar } from 'libs/hooks/useWhiteStatusBar'
-import { WebShareModal } from 'libs/share/WebShareModal'
 import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
 import { RoundedButton } from 'ui/components/buttons/RoundedButton'
 import { useModal } from 'ui/components/modals/useModal'
